Rename `components/snackbar/` -> `components/alertsnackbar/`
